fix: did you say grub support?

+++ src/kernel/runtime/stack_protector.c
@@ -3,7 +3,7 @@
 #include <stdint.h>
 #include <stdnoreturn.h>
 
-uint64_t __stack_chk_guard = (uint64_t)__TIMESTAMP__;
+size_t __stack_chk_guard = (size_t)__TIMESTAMP__;
 
 noreturn void __stack_chk_fail(void) {
     error$("stack protector check fail");
